requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aenum==3.1.15
    # via line-bot-sdk
aiofiles==23.2.1
    # via gradio
aiohttp==3.9.3
    # via line-bot-sdk
aiosignal==1.3.1
    # via aiohttp
altair==5.3.0
    # via gradio
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
boto3==1.34.80
botocore==1.34.80
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deprecated==1.2.14
    # via line-bot-sdk
fastapi==0.110.1
    # via gradio
ffmpy==0.3.2
    # via gradio
filelock==3.13.3
    # via huggingface-hub
fonttools==4.51.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.3.1
    # via
    #   gradio-client
    #   huggingface-hub
future==1.0.0
    # via line-bot-sdk
gradio==4.26.0
gradio-client==0.15.1
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
httpx==0.27.0
    # via
    #   gradio
    #   gradio-client
huggingface-hub==0.22.2
    # via
    #   gradio
    #   gradio-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.4.0
    # via gradio
jinja2==3.1.3
    # via
    #   altair
    #   gradio
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
line-bot-sdk==3.10.1
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.8.4
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via
    #   altair
    #   contourpy
    #   gradio
    #   matplotlib
    #   pandas
orjson==3.10.0
    # via gradio
packaging==24.0
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   matplotlib
pandas==2.2.1
    # via
    #   altair
    #   gradio
pillow==10.3.0
    # via
    #   gradio
    #   matplotlib
pydantic==2.6.4
    # via
    #   fastapi
    #   gradio
    #   line-bot-sdk
pydantic-core==2.16.3
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.17.2
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   line-bot-sdk
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
python-multipart==0.0.9
    # via gradio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   gradio
    #   huggingface-hub
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   huggingface-hub
    #   line-bot-sdk
rich==13.7.1
    # via typer
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
ruff==0.3.5
    # via gradio
s3transfer==0.10.1
    # via boto3
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tomlkit==0.12.0
    # via gradio
toolz==0.12.1
    # via altair
tqdm==4.66.2
    # via huggingface-hub
typer==0.12.2
    # via gradio
typing-extensions==4.11.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via
    #   botocore
    #   line-bot-sdk
    #   requests
uvicorn==0.29.0
    # via gradio
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
    # via gradio-client
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via aiohttp
